requirements.txt
opencv-python
numpy
pillow
mediapipe
msvc-runtime
PyYAML
